+++ src/routes/commentRoutes.js
@@ -8,6 +8,7 @@ commentRoutes
  .post('/api/comments', bearer, CommentController.createComment)
  .get('/api/:divingSpotId/comments', CommentController.findCommentsByDivingSpotId)
  .get('/api/comments/:id', CommentController.findCommentById)
+ .get('/api/commentsByUserToken', CommentController.findCommentsByUserToken)
  .put('/api/comments/:id', bearer, CommentController.updateComment)
  .delete('/api/comments/:id', bearer, CommentController.deleteComment);