+++ CHANGELOG.md
@@ -2,6 +2,49 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2.0.0] - 2024-06-02
+
+### Added
+- DiveLogs routes:
+    - /api/diveLogs/date
+    - /api/diveLogs/location/:locationName
+- Comment routes:
+    - /api/commentsByUserToken
+- logger;
+- MongoDB Atlas.
+
+### Changed
+- Create dynamic statistics based on the period the user requests;
+- Update libs:
+   - @sendgrid/mail to v8.1.3,
+   - bcryptjs to v2.4.3,
+   - cors to v2.8.5,
+   - dotenv  to v16.4.5,
+   - express to v4.19.2,
+   - joi to v17.13.1,
+   - jsonwebtoken to v9.0.2,
+   - mongoose to v7.6.12,
+   - passport to v0.7.0,
+   - passport-http-bearer to v1.0.1,
+   - passport-local to v1.0.0,
+   - swagger-ui-express to v4.6.3,
+   - winston to v3.13.0,
+   - yaml to v2.4.2,
+   - jest to v29.7.0,
+   - nodemon to v3.1.2,
+   - supertest to v6.3.4.
+
+### Removed
+- diveStatistics Model;
+- diveStatistics Repository;
+- DiveLogs routes:
+    - /api/diveLogs/rating/:rating
+    - /api/diveLogs/location/:location
+- users route:
+    - /api/users
+- docker;
+- eslint.
+
 ## [1.10.0] - 2024-05-04
 
 ### Added
@@ -20,7 +63,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.8.1] - 2024-04-08
 
-### Adjusted
+### Changed
 - Statistics calculation, as WeatherCondition and WeatherBody can be null;
 - Swagger, id is not in uuid format.
 
@@ -42,7 +85,7 @@ All notable changes to this project will be documented in this file.
 ### Removed
 - docker-compose.debug and docker-compose.environment.
 
-### Adjusted
+### Changed
 - Refactoring Controller and Service.
 
 ## [1.6.0] - 2024-04-02